+++ pyxem/data/__init__.py
@@ -30,7 +30,14 @@ then be deleted manually if desired.
 """
 
 from pyxem.data.simulated_tilt import tilt_boundary_data
-from pyxem.data.simulated_dpc import simulated_stripes
+from pyxem.data.simulated_si import (
+    si_phase,
+    si_tilt,
+    si_grains,
+    si_grains_simple,
+    si_rotations_line,
+)
+from pyxem.data.simulated_fe import fe_bcc_phase, fe_fcc_phase, fe_multi_phase_grains
 from pyxem.data._data import (
     au_grating,
     pdnip_glass,
@@ -38,6 +45,11 @@ from pyxem.data._data import (
     twinned_nanowire,
     sample_with_g,
     mgo_nanocrystals,
+    organic_semiconductor,
+    cuag_orientations,
+    feal_stripes,
+    sped_ag,
+    pdcusi_insitu,
 )
 
 __all__ = [
@@ -48,5 +60,18 @@ __all__ = [
     "sample_with_g",
     "mgo_nanocrystals",
     "tilt_boundary_data",
-    "simulated_stripes",
+    "si_phase",
+    "si_tilt",
+    "si_grains",
+    "si_grains_simple",
+    "si_rotations_line",
+    "fe_multi_phase_grains",
+    "fe_fcc_phase",
+    "fe_bcc_phase",
+    "cuag_orientations",
+    "organic_semiconductor",
+    "feal_stripes",
+    "sped_ag",
+    "pdcusi_insitu",
+>>>>>>> main
 ]